requirements.txt
timm
diffusers
einops
torchprofile
accelerate
tensorflow
tqdm
scipy
requests
wandb
matplotlib